+++ __tests__/BootstrapTable-test.js
@@ -7,7 +7,7 @@ jest.dontMock('../src/TableRow.js');
 jest.dontMock('../src/pagination/PaginationList.js');
 jest.dontMock('../src/pagination/PageButton.js');
 
-describe('BootstrapTable', function() {
+xdescribe('BootstrapTable', function() {
   var testData = [
     {id: 1, name: "name1", price: 100},
     {id: 2, name: "name2", price: 120},